refactor: consolidate language maps, batched SQL, AsyncOpQueue, bruteForceVectorSearch
- Task 5: Move EXT_TO_LANG, EXT_TO_LANGUAGE_NAME, EXT_TO_DYNAMIC_LANG, LANGUAGE_EXTENSIONS to ast-languages.ts
- Task 6: Extract runBatchedQuery helper in store.ts, refactor 5 batched methods
- Task 7: Extract AsyncOpQueue to utils/async.ts, update lancedb + qdrant stores
- Task 8: Extract bruteForceVectorSearch to vector-store.ts, add removeMany to interface
